adding policies and addition functions

+++ app/Models/Payments/ClientPayment.php
@@ -49,6 +49,10 @@ class ClientPayment extends Model
     ///model functions
     public function setInfo(Carbon $due, $type, $note = null)
     {
+        /** @var User */
+        $user = Auth::user();
+        if (!$user->can('update', $this)) return false;
+
         try {
             AppLog::error("Setting Client Payment info", loggable: $this);
             return $this->update([
@@ -64,6 +68,10 @@ class ClientPayment extends Model
 
     public function setDocument($doc_url)
     {
+        /** @var User */
+        $user = Auth::user();
+        if (!$user->can('update', $this)) return false;
+
         try {
             if ($this->doc_url)
                 Storage::delete($this->doc_url);
@@ -81,6 +89,10 @@ class ClientPayment extends Model
 
     public function deleteDocument()
     {
+        /** @var User */
+        $user = Auth::user();
+        if (!$user->can('update', $this)) return false;
+
         try {
             if ($this->doc_url)
                 Storage::delete($this->doc_url);
@@ -95,15 +107,19 @@ class ClientPayment extends Model
 
     public function setAsPaid(Carbon $date = null)
     {
+        /** @var User */
+        $user = Auth::user();
+        if (!$user->can('update', $this)) return false;
+
         if (!$this->is_new) return false;
         try {
             $date = $date ?? new Carbon();
             AppLog::error("Setting Client Payment as paid", loggable: $this);
-            if($this->update([
+            if ($this->update([
                 "closed_by_id"   =>  Auth::id(),
                 "payment_date"  => $date->format('Y-m-d H:i'),
                 "status"  =>  self::PYMT_STATE_PAID,
-            ])){
+            ])) {
                 $this->loadMissing('sold_policy');
                 $this->sold_policy->setClientPaymentDate($date);
                 $this->sold_policy->calculateTotalClientPayments();
@@ -117,6 +133,10 @@ class ClientPayment extends Model
 
     public function setAsCancelled(Carbon $date = null)
     {
+        /** @var User */
+        $user = Auth::user();
+        if (!$user->can('update', $this)) return false;
+
         if (!$this->is_new) return false;
         try {
             $date = $date ?? new Carbon();
